+++ src/ZenGTPX/Gtp/IGtpEngine.cs
@@ -12,11 +12,19 @@ public interface IGtpEngine
 
     void SetKomi(double komi);
 
+    void SetNextColor(StoneColor color);
+
     void SetMaxTime(double seconds);
+
+    void SetTimeSettings(double mainTime, double byoyomiTime, int periods);
+
+    void SetTimeLeft(StoneColor color, double time, int stones);
 
     bool Play(StoneColor color, GtpMove move);
 
     GtpMove GenMove(StoneColor color);
 
     bool Undo(int count);
+
+    string EstimateFinalScore();
 }